Fixed bug "Fullscreen overlay does not resize on orientation changed" (https://github.com/jorgesilva/TeclaAccess/issues/14).
Problem was caused by the fact that the fullscreen switch overlay
didn't change size/shape when the screen orientation changed. To fix this,
I've overridden onConfigurationChanged (which triggers when the orientation
changes) to send out a new intent, ACTION_RESTART_FS_SWITCH_MODE. This works
similarly to the ACTION_START_FS_SWITCH_MODE intent except that it won't
show the "Fullscreen switch enabled" toast.

To toggle the Toast, I modified mCreateSwitchRunnable to be an instance of
an inner class that takes a boolean argument, showToast. startFullScreenSwitchMode()
provides the boolean.